fix Post, comment, Reaction

+++ app/Http/Controllers/Api/commentcontroller.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\ShowUserNametPostedResource;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 
@@ -30,8 +31,13 @@ class commentcontroller extends Controller
     public function update(Request $request, string $id)
     {
         //
+        $comment = Comment::find($id);
+        if (  Auth()-> user() -> id  == $comment-> user_id){
         Comment::updateComment($request, $id);
-        return ["success" => true, "Message" =>"comment updated successfully"];
+        return ["success" => true, "Message" =>"comment updated successfully", "User_Comment" => new ShowUserNametPostedResource($comment), "User_Reaction" => Auth()-> user() -> name];
+        }else{
+            return ["success" => false, "Message" =>"you are not allowed to update this comment" , "User_Comment" => new ShowUserNametPostedResource($comment), "User_update_Reaction" => Auth()-> user() -> name];
+        }
     }
 
     /**
@@ -40,7 +46,12 @@ class commentcontroller extends Controller
     public function destroy(string $id)
     {
         //
+        $comment = Comment::find($id);
+        if (  Auth()-> user() -> id  == $comment-> user_id){
         Comment::deleteComment($id);
-        return ["success" => true, "Message" =>"comment deleted successfully"];
-    }
+        return ["success" => true, "Message" =>"comment deleted successfully", "User_Commenat" => new ShowUserNametPostedResource($comment), "User_Reaction" => Auth()-> user() -> name];
+        }else{
+            return ["success" => false, "Message" =>"you are not allowed to delete this comment" , "User_Comment" => new ShowUserNametPostedResource($comment), "User_update_Reaction" => Auth()-> user() -> name];
+    
+    }}
 }
